+++ stroom-explorer-server/src/main/java/stroom/explorer/server/ExplorerSession.java
@@ -1,0 +1,9 @@
+package stroom.explorer.server;
+
+import java.util.Optional;
+
+interface ExplorerSession {
+    Optional<Long> getMinExplorerTreeModelBuildTime();
+
+    void setMinExplorerTreeModelBuildTime(long buildTime);
+}